Update pgRouting-concepts.rst (#1876)
Added a couple of paragraph explaining the use of aliases in queries in pgRouting concepts section. Added a query example that should work out of the box for those using osm2pgrouting to import OSM data.

This is one part of the documentation that is first read by new users and all the basic concepts should be introduced there, including the options to use databases with different structures and how to use aliases. Also, all the samples should work out of the box when using osm2pgrouting-generated data.

This pull request should fix the urgency in osm2pgrouting's issue #290 and discussion can now continue without short-term pressure.